+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.91.43] - 2025-08-03 19:00:00
+### Added
+- Support swatch type attributes for products and variants
+- Support multiple images per variant
+
 ## [4.91.42] - 2025-07-27 20:00:00
 ### Fixed
 - Adhere to PHP 8.4 standards by adding nullable type hints in AbstractCollection constructor